reduce removed style with INHERITABLE_PROPERTIES
