+++ packages/roosterjs-editor-dom/lib/htmlSanitizer/getInheritableStyles.ts
@@ -2,7 +2,7 @@ import type { StringMap } from 'roosterjs-editor-types';
 
 // Inheritable CSS properties
 // Ref: https://www.w3.org/TR/CSS21/propidx.html
-const INHERITABLE_PROPERTIES = (
+export const INHERITABLE_PROPERTIES = (
     'border-spacing,caption-side,color,' +
     'cursor,direction,empty-cells,font-family,font-size,font-style,font-variant,font-weight,' +
     'font,letter-spacing,line-height,list-style-image,list-style-position,list-style-type,' +
